refactor : MVC 패턴에 맞추어서 패키지별로 클래스 구성. 사용자 입력값에 대해 유효성 검사하는 부분을 별도의 클래스로 분리
