+++ src/test/java/carRace/utils/RandomGeneratorTest.java
@@ -1,0 +1,4 @@
+package carRace.utils;
+
+public class RandomGeneratorTest {
+}
