+++ src/test/java/carRace/utils/StringUtilTest.java
@@ -1,0 +1,4 @@
+package carRace.utils;
+
+public class StringUtilTest {
+}
